// src/types/styles.d.ts
// types/styles.d.ts

// Define the shape types
export type Shape = 'circle' | 'ellipse' | 'rectangle' | 'roundedRectangle' | 'equilateralTriangle';

// Define the text position types
export type TextPosition = 
  | 'topLeft'
  | 'topCenter'
  | 'topRight'
  | 'centerLeft'
  | 'center'
  | 'centerRight'
  | 'bottomLeft'
  | 'bottomCenter'
  | 'bottomRight';